+++ modules/tinymce/src/plugins/media/main/ts/core/Nodes.ts
@@ -11,9 +11,9 @@ import Editor from 'tinymce/core/api/Editor';
 import Env from 'tinymce/core/api/Env';
 import DomParser from 'tinymce/core/api/html/DomParser';
 import AstNode from 'tinymce/core/api/html/Node';
+import HtmlSerializer from 'tinymce/core/api/html/Serializer';
 
 import * as Options from '../api/Options';
-import * as Sanitize from './Sanitize';
 
 declare let escape: any;
 
@@ -59,7 +59,6 @@ const createPlaceholderNode = (editor: Editor, node: AstNode): AstNode => {
   const name = node.name;
 
   const placeHolder = new AstNode('img', 1);
-  placeHolder.shortEnded = true;
 
   retainAttributesAndInnerHtml(editor, node, placeHolder);
 
@@ -144,10 +143,13 @@ const retainAttributesAndInnerHtml = (editor: Editor, sourceNode: AstNode, targe
 
   // Place the inner HTML contents inside an escaped attribute
   // This enables us to copy/paste the fake object
-  const innerHtml = sourceNode.firstChild && sourceNode.firstChild.value;
+  const serializer = HtmlSerializer({ inner: true }, editor.schema);
+  const tempNode = new AstNode('div', 1);
+  Arr.each(sourceNode.children(), (child) => tempNode.append(child));
+  const innerHtml = serializer.serialize(tempNode);
   if (innerHtml) {
-    targetNode.attr('data-mce-html', escape(Sanitize.sanitize(editor, innerHtml)));
-    targetNode.firstChild = null;
+    targetNode.attr('data-mce-html', escape(innerHtml));
+    targetNode.empty();
   }
 };